skeleton/backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
pyxlsb
langchain-google-genai
python-dotenv
watchdog
